requirements.txt
pandas==0.22.0
matplotlib==2.2.2
scikit-learn==0.19.1
scipy==1.0.1
cherrypy==14.2.0
pillow==5.1.0
